requirements.txt
django-singleton
